added setting_row_label component
